+++ include/sql/sample/_develop.sql
@@ -101,7 +101,8 @@ INSERT INTO person_info_opt_field
 ('sports', 'ja',           'スポーツ', 'text',        1);
 
 -- 管理画面メニューデータ
-DELETE FROM _nav_item;
+DELETE FROM _nav_item WHERE ni_nav_id = 'admin_menu';
+DELETE FROM _nav_item WHERE ni_nav_id = 'admin_menu.en';
 INSERT INTO _nav_item
 (ni_id, ni_parent_id, ni_index, ni_nav_id,    ni_task_id,        ni_view_control, ni_param, ni_name,                ni_help_title,          ni_help_body) VALUES
 (100,   0,            0,        'admin_menu', '_page',           0,               '',       '画面管理',             '画面管理',             'Webサイトのデザインや機能を管理します。'),
